requirements.txt
pyparsing==2.3.0
pysvg-py3==0.2.2.post2
svgwrite==1.2.1
